updating README to include steps to run metrics locally. separating the metrics into the controller (for endpoint) and the service (for middleware work)
