+++ app/controllers/hello_controller.py
@@ -2,12 +2,12 @@ from fastapi import APIRouter
 
 from app.models.hello_dto import HelloRequest
 
-router = APIRouter()
+hello_router = APIRouter()
 
-@router.get("/hello")
+@hello_router.get("/hello")
 def hello():
     return {"message": "Hello, world!"}
 
-@router.post("/hello")
+@hello_router.post("/hello")
 def hello_post(request: HelloRequest):
     return {"message": f"Hello, {request.name}!"}
