+++ app/controllers/metrics_controller.py
@@ -1,0 +1,10 @@
+from fastapi import APIRouter
+from prometheus_client import generate_latest, CONTENT_TYPE_LATEST
+from starlette.responses import Response
+
+metrics_router = APIRouter()
+
+@metrics_router.get("/metrics")
+async def metrics():
+    data = generate_latest()
+    return Response(content=data, media_type=CONTENT_TYPE_LATEST)
