+++ app/main.py
@@ -1,10 +1,11 @@
 from fastapi import FastAPI
-from app.controllers.hello_controller import router as api_router
-from app.metrics import metrics_middleware, metrics_router
+from app.controllers.hello_controller import hello_router
+from app.controllers.metrics_controller import metrics_router
+from app.service.metrics_service import metrics_middleware
 
 app = FastAPI(title="FastAPI Skeleton")
 
-app.include_router(api_router, prefix="/v1")
+app.include_router(hello_router, prefix="/v1")
 app.include_router(metrics_router)
 
 app.middleware("http")(metrics_middleware)
